@@ -1,6 +1,6 @@
 {
   "name": "icon-with-tooltip",
-  "version": "1.5.1",
+  "version": "2.0.0",
   "description": "An icon displaying a tooltip on hover.",
   "main": "lib/index.js",
   "sideEffects": false,
@@ -11,7 +11,8 @@
     "storybook:start": "start-storybook -c storybook",
     "storybook:build": "build-storybook -c storybook -o docs",
     "storybook:commit": "git add . && git commit -m \"Updated storybook\"",
-    "all": "yarn install && yarn storybook:build && yarn deploy"
+    "all": "yarn reinstall && yarn storybook:build && yarn deploy",
+    "reinstall": "rimraf yarn.lock node_modules && yarn install"
   },
   "repository": {
     "type": "git",
@@ -37,43 +38,29 @@
   },
   "homepage": "https://github.com/iulian-radu-at/icon-with-tooltip#readme",
   "dependencies": {
-    "@material-ui/core": "^4.11.3",
-    "@material-ui/icons": "^4.11.2"
+    "@mui/icons-material": "^5.5.1",
+    "@mui/material": "^5.5.1",
+    "@mui/styles": "^5.5.1"
   },
   "devDependencies": {
-    "@babel/core": "^7.13.8",
-    "@storybook/react": "6.0.28",
-    "@types/babel-core": "^6.25.6",
-    "@types/enzyme": "^3.10.8",
-    "@types/enzyme-adapter-react-16": "^1.0.6",
-    "@types/react": "^16.9.56",
-    "@types/react-dom": "^16.9.11",
-    "@types/rimraf": "^3.0.0",
+    "@babel/core": "^7.17.8",
+    "@emotion/react": "^11.8.2",
+    "@storybook/react": "^6.4.19",
+    "@types/babel-core": "^6.25.7",
+    "@types/react": "^17.0.41",
+    "@types/react-dom": "^17.0.14",
+    "@types/rimraf": "^3.0.2",
     "babel-core": "^6.26.3",
-    "babel-jest": "^26.6.3",
-    "babel-loader": "^8.2.2",
-    "babel-preset-react-app": "^10.0.0",
+    "babel-loader": "^8.2.3",
+    "babel-preset-react-app": "^10.0.1",
     "cross-env": "^7.0.3",
-    "enzyme": "^3.11.0",
-    "enzyme-adapter-react-16": "^1.15.6",
+    "react": "^17.0.2",
+    "react-dom": "^17.0.2",
     "rimraf": "^3.0.2",
-    "typescript": "^4.2.2"
+    "typescript": "^4.6.2"
   },
   "peerDependencies": {
-    "react": ">= 16.9.0",
-    "react-dom": ">= 16.9.0"
-  },
-  "jest": {
-    "modulePathIgnorePatterns": [
-      "./node_modules",
-      "./history"
-    ],
-    "transform": {
-      "^.+\\.js$": "babel-jest"
-    },
-    "testRegex": "src/*(/(__tests?__/)([^_].*/)*?[^_][^/]*?\\.(test|spec)?\\.(js?))$",
-    "setupFiles": [
-      "./enzymeAdapter.setup.js"
-    ]
+    "react": ">=17.0.0",
+    "react-dom": ">=17.0.0"
   }
 }
