@@ -1,43 +1,47 @@
-import * as React from "react";
-import { IconWithTooltipProps } from "./IconWithTooltip";
-import { StyledComponentProps, Tooltip } from "@material-ui/core";
-import { withStyles } from "@material-ui/styles";
+import { Tooltip } from '@mui/material';
+import { makeStyles } from '@mui/styles';
+import * as React from 'react';
+import { IconWithTooltipProps } from './IconWithTooltip';
 
-const stylesTooltip = {
+const useStyles = makeStyles({
   tooltip: {
-    color: "#222222",
-    backgroundColor: "#dddddd",
-    margin: "2px 5px",
+    color: '#222222',
+    backgroundColor: '#dddddd',
+    margin: '2px 5px',
   },
-};
+});
 
 const StyledTooltip: React.FC<StyledTooltipProps> = ({
-  classes,
   Icon,
   placement,
   text,
-}) => (
-  <Tooltip title={text || ""} placement={placement} classes={classes}>
-    <Icon />
-  </Tooltip>
-);
+}) => {
+  const classes = useStyles();
+  if (!text) {
+    return <Icon />;
+  }
+
+  return (
+    <Tooltip title={text} placement={placement} classes={classes}>
+      <Icon />
+    </Tooltip>
+  );
+};
 
 export type TooltipPlacement =
-  | "bottom-end"
-  | "bottom-start"
-  | "bottom"
-  | "left-end"
-  | "left-start"
-  | "left"
-  | "right-end"
-  | "right-start"
-  | "right"
-  | "top-end"
-  | "top-start"
-  | "top";
+  | 'bottom-end'
+  | 'bottom-start'
+  | 'bottom'
+  | 'left-end'
+  | 'left-start'
+  | 'left'
+  | 'right-end'
+  | 'right-start'
+  | 'right'
+  | 'top-end'
+  | 'top-start'
+  | 'top';
 
-interface StyledTooltipProps extends IconWithTooltipProps {
-  classes: Record<"tooltip", string>;
-}
+interface StyledTooltipProps extends IconWithTooltipProps {}
 
-export default withStyles(stylesTooltip)(StyledTooltip);
+export default StyledTooltip;
