@@ -1,14 +1,17 @@
+import InfoIcon from '@mui/icons-material/Info';
 import * as React from 'react';
 import IconWithTooltip from './IconWithTooltip';
-import InfoIcon from '@material-ui/icons/Info';
 import { TooltipPlacement } from './StyledTooltip';
 
 const style: React.CSSProperties = {
   color: '#0040ff',
-  display: 'inline-block'
+  display: 'inline-block',
 };
 
-export const InfoIconWithTooltip: React.FC<InfoIconWithTooltipProps> = ({ placement, text }): JSX.Element => (
+export const InfoIconWithTooltip: React.FC<InfoIconWithTooltipProps> = ({
+  placement,
+  text,
+}): JSX.Element => (
   <div style={style}>
     <IconWithTooltip Icon={InfoIcon} text={text} placement={placement} />
   </div>
